Improve robustness of resuming unfinished extrap
Write out a status file for each time chunk and step of the process.
Once a step is finished, it is logged.  When resuming, only redo
any unfinished steps.

We cannot use output files to determine if a process is done
because they exist from the start of the step.